Federal Papers Analysis using Decision Trees
Federal Papers written by Jay , Hamilton, and Madison. Several of the papers have unknown authors. Here, EDA is performed on the data to get insights into the writing styles of each of the authors. After the EDA, Decision Trees are used in order to try and determine the most likely author of the unknown papers.